services/data_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pandas==2.1.3
numpy==1.24.3
boto3==1.29.3
python-dotenv==1.0.0
python-multipart==0.0.6 